requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audeer==2.0.0
audiofile==1.4.0
audioread==3.0.1
audmath==1.4.0
audtorch==0.6.4
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.3.3
causal-conv1d==1.0.0
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi==1.16.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
einops==0.8.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
huggingface-hub==0.23.0
idna @ file:///croot/idna_1714398848350/work
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.3.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning-utilities==0.11.2
llvmlite==0.42.0
mamba-ssm==1.0.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
ml-dtypes==0.3.2
mne==1.7.0
msgpack==1.0.8
multidict==6.0.5
namex==0.0.8
ninja==1.11.1.1
numba==0.59.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=6094eeedd869502faa0fd0a8c5ad3a70c5779be06ddd1feb7627e5c212fac420
nvidia-cuda-runtime-cu12==12.4.127
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pillow @ file:///croot/pillow_1714398848491/work
platformdirs==4.2.2
pooch==1.8.1
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.10
requests @ file:///croot/requests_1707355572290/work
requests-oauthlib==2.0.0
resampy==0.4.3
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorboard_logger==0.0.1
tensorflow==2.16.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.37.0
tensorrt==10.0.1
tensorrt-cu12==10.0.1
tensorrt-cu12-bindings==10.0.1
tensorrt-cu12-libs==10.0.1
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.1
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.4.0.post0
torchvision==0.14.1
tqdm==4.66.4
transformers==4.40.2
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2024.1
urllib3 @ file:///croot/urllib3_1715635851070/work
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
